bug2

+++ src/main/java/it/polimi/ingsw/GC_23/Connection/MatchTimeOut.java
@@ -10,7 +10,7 @@ import java.util.logging.Logger;
  * Created by jesss on 16/06/17.
  */
 public class MatchTimeOut implements Runnable {
-    private long time = ParseJson.getParseJson().getTimeoutStartMatch() * 1000;
+    private int time = ParseJson.getParseJson().getTimeoutStartMatch() * 1000;
     private Match match;
     private final Logger logger = Logger.getLogger(Logger.GLOBAL_LOGGER_NAME);
